+++ Gateway/Config/HgwBasePaymentConfig.php
@@ -46,11 +46,12 @@ class HgwBasePaymentConfig extends BaseConfig implements HgwBasePaymentConfigInt
     /**
      * Get payment channel id.
      *
+     * @param int|null $storeId
      * @return string
      */
-    public function getChannel()
+    public function getChannel($storeId = null)
     {
-        return $this->getValue(self::KEY_CHANNEL);
+        return $this->getValue(self::KEY_CHANNEL, $storeId);
     }
 
     /**